backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
firebase-admin==6.5.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.3.0
google-generativeai==0.8.2
redis==5.1.1
python-multipart==0.0.12
pdfplumber==0.11.4
python-docx==1.1.2
Jinja2==3.1.4
aiofiles==24.1.0
httpx==0.27.2
requests==2.31.0
pytest==8.3.3
pytest-asyncio==0.24.0
